feat(admin): add period navigation to stats pages
