+++ apps/admin/src/app/(private)/stats/_components/admin-period-navigation.tsx
@@ -1,0 +1,48 @@
+"use client";
+
+import { Button } from "@zoonk/ui/components/button";
+import { ChevronLeftIcon, ChevronRightIcon } from "lucide-react";
+import { useQueryState } from "nuqs";
+
+export function AdminPeriodNavigation({
+  hasNext,
+  periodLabel,
+}: {
+  hasNext: boolean;
+  periodLabel: string;
+}) {
+  const [offset, setOffset] = useQueryState("offset", {
+    defaultValue: "0",
+    shallow: false,
+  });
+
+  const currentOffset = Number(offset) || 0;
+
+  function handlePrevious() {
+    void setOffset(String(currentOffset + 1));
+  }
+
+  function handleNext() {
+    void setOffset(String(Math.max(0, currentOffset - 1)));
+  }
+
+  return (
+    <div className="flex items-center gap-2">
+      <Button aria-label="Previous period" onClick={handlePrevious} size="icon" variant="outline">
+        <ChevronLeftIcon aria-hidden />
+      </Button>
+
+      <span className="min-w-32 text-center text-sm font-medium">{periodLabel}</span>
+
+      <Button
+        aria-label="Next period"
+        disabled={!hasNext}
+        onClick={handleNext}
+        size="icon"
+        variant="outline"
+      >
+        <ChevronRightIcon aria-hidden />
+      </Button>
+    </div>
+  );
+}
