+++ apps/admin/src/app/(private)/stats/content/page.tsx
@@ -1,5 +1,7 @@
+import { calculateDateRanges, formatPeriodLabel, validatePeriod } from "@zoonk/utils/date-ranges";
 import { type Metadata } from "next";
 import { Suspense } from "react";
+import { AdminPeriodNavigation } from "../_components/admin-period-navigation";
 import { StatsPageLayout } from "../_components/stats-page-layout";
 import { ContentMetrics, ContentMetricsSkeleton } from "./content-metrics";
 
@@ -8,11 +10,22 @@ export const metadata: Metadata = {
 };
 
 export default async function ContentPage({ searchParams }: PageProps<"/stats/content">) {
-  const { period } = await searchParams;
+  const { period: rawPeriod, offset: rawOffset } = await searchParams;
+  const period = validatePeriod(String(rawPeriod ?? "month"));
+  const offset = Number(rawOffset) || 0;
+  const { current } = calculateDateRanges(period, offset);
+  const periodLabel = formatPeriodLabel(current.start, current.end, period, "en");
 
   return (
-    <StatsPageLayout title="Content & Operations">
-      <Suspense fallback={<ContentMetricsSkeleton />} key={String(period)}>
+    <StatsPageLayout
+      navigation={
+        period === "all" ? null : (
+          <AdminPeriodNavigation hasNext={offset > 0} periodLabel={periodLabel} />
+        )
+      }
+      title="Content & Operations"
+    >
+      <Suspense fallback={<ContentMetricsSkeleton />} key={`${period}-${offset}`}>
         <ContentMetrics searchParams={searchParams} />
       </Suspense>
     </StatsPageLayout>
