+++ apps/admin/src/app/(private)/stats/growth/growth-metrics.tsx
@@ -13,11 +13,12 @@ import { SubscribersTable } from "./subscribers-table";
 export async function GrowthMetrics({
   searchParams,
 }: {
-  searchParams: Promise<{ period?: string }>;
+  searchParams: Promise<{ period?: string; offset?: string }>;
 }) {
-  const { period: rawPeriod } = await searchParams;
+  const { period: rawPeriod, offset: rawOffset } = await searchParams;
   const period = validatePeriod(rawPeriod ?? "month");
-  const { current, previous } = calculateDateRanges(period, 0);
+  const offset = Number(rawOffset) || 0;
+  const { current, previous } = calculateDateRanges(period, offset);
 
   const [
     currentSignups,
@@ -43,7 +44,7 @@ export async function GrowthMetrics({
 
   return (
     <div className="flex flex-col gap-8">
-      <div className="grid grid-cols-1 gap-x-4 gap-y-6 sm:grid-cols-2 lg:grid-cols-3">
+      <div className="grid grid-cols-1 gap-x-8 gap-y-6 sm:grid-cols-2 lg:grid-cols-3">
         <AdminMetricCard
           change={{ current: currentSignups, period, previous: previousSignups }}
           help="New user registrations in this period"
@@ -74,8 +75,12 @@ export async function GrowthMetrics({
       )}
 
       {subscribers.length > 0 && (
-        <div className="rounded-lg border">
-          <SubscribersTable data={subscribers} />
+        <div className="flex flex-col gap-3">
+          <h3 className="text-base font-semibold tracking-tight">Subscribers by Plan</h3>
+
+          <div className="rounded-lg border">
+            <SubscribersTable data={subscribers} />
+          </div>
         </div>
       )}
     </div>
@@ -85,7 +90,7 @@ export async function GrowthMetrics({
 export function GrowthMetricsSkeleton() {
   return (
     <div className="flex flex-col gap-8">
-      <div className="grid grid-cols-1 gap-x-4 gap-y-6 sm:grid-cols-2 lg:grid-cols-3">
+      <div className="grid grid-cols-1 gap-x-8 gap-y-6 sm:grid-cols-2 lg:grid-cols-3">
         <AdminMetricCardSkeleton />
         <AdminMetricCardSkeleton />
         <AdminMetricCardSkeleton />
