+++ apps/admin/src/app/(private)/stats/growth/page.tsx
@@ -1,5 +1,7 @@
+import { calculateDateRanges, formatPeriodLabel, validatePeriod } from "@zoonk/utils/date-ranges";
 import { type Metadata } from "next";
 import { Suspense } from "react";
+import { AdminPeriodNavigation } from "../_components/admin-period-navigation";
 import { StatsPageLayout } from "../_components/stats-page-layout";
 import { GrowthMetrics, GrowthMetricsSkeleton } from "./growth-metrics";
 
@@ -8,11 +10,22 @@ export const metadata: Metadata = {
 };
 
 export default async function GrowthPage({ searchParams }: PageProps<"/stats/growth">) {
-  const { period } = await searchParams;
+  const { period: rawPeriod, offset: rawOffset } = await searchParams;
+  const period = validatePeriod(String(rawPeriod ?? "month"));
+  const offset = Number(rawOffset) || 0;
+  const { current } = calculateDateRanges(period, offset);
+  const periodLabel = formatPeriodLabel(current.start, current.end, period, "en");
 
   return (
-    <StatsPageLayout title="Growth & Sustainability">
-      <Suspense fallback={<GrowthMetricsSkeleton />} key={String(period)}>
+    <StatsPageLayout
+      navigation={
+        period === "all" ? null : (
+          <AdminPeriodNavigation hasNext={offset > 0} periodLabel={periodLabel} />
+        )
+      }
+      title="Growth & Sustainability"
+    >
+      <Suspense fallback={<GrowthMetricsSkeleton />} key={`${period}-${offset}`}>
         <GrowthMetrics searchParams={searchParams} />
       </Suspense>
     </StatsPageLayout>
