--- apps/main/src/app/(performance)/_utils.ts
@@ -1,33 +1,0 @@
-import { type HistoryPeriod } from "@zoonk/utils/date-ranges";
-
-export function formatPeriodLabel(
-  periodStart: Date,
-  periodEnd: Date,
-  period: HistoryPeriod,
-  locale: string,
-): string {
-  if (period === "all") {
-    return `${periodStart.getFullYear()} - ${periodEnd.getFullYear()}`;
-  }
-
-  if (period === "month") {
-    return new Intl.DateTimeFormat(locale, {
-      month: "long",
-      year: "numeric",
-    }).format(periodStart);
-  }
-
-  if (period === "6months") {
-    const startMonth = new Intl.DateTimeFormat(locale, {
-      month: "short",
-    }).format(periodStart);
-
-    const endMonth = new Intl.DateTimeFormat(locale, { month: "short" }).format(periodEnd);
-
-    const year = periodStart.getFullYear();
-
-    return `${startMonth} - ${endMonth} ${year}`;
-  }
-
-  return String(periodStart.getFullYear());
-}
