+++ apps/main/src/app/(performance)/score/score-stats.tsx
@@ -1,8 +1,7 @@
 import { Skeleton } from "@zoonk/ui/components/skeleton";
-import { type HistoryPeriod } from "@zoonk/utils/date-ranges";
+import { type HistoryPeriod, formatPeriodLabel } from "@zoonk/utils/date-ranges";
 import { getExtracted, getLocale } from "next-intl/server";
 import { MetricComparison } from "../_components/metric-comparison";
-import { formatPeriodLabel } from "../_utils";
 
 export async function ScoreStats({
   average,
