+++ apps/main/src/data/progress/get-best-time.ts
@@ -1,8 +1,9 @@
 import "server-only";
 import { getSession } from "@zoonk/core/users/session/get";
 import { prisma } from "@zoonk/db";
-import { type ScoredRow, findBestByScore, getDefaultStartDate } from "@zoonk/utils/date-ranges";
+import { getDefaultStartDate } from "@zoonk/utils/date-ranges";
 import { safeAsync } from "@zoonk/utils/error";
+import { type ScoredRow, findBestByScore } from "@zoonk/utils/score-aggregation";
 import { cache } from "react";
 
 type BestTimeData = {
